chore: combines update and write env functions
